First real commit...
Brought all the various pieces together into this, their proper metadata
layer.

- Abseil for TF-Lite.
- TF-Lite for Python3.
- TF-Lite for C++.
- Google Gasket drivers for Gasket proper and the EdgeTPU driver.
- Edge TPU support libraries for those drivers.

To be sure, there's going to be some updates to some of this very
shortly because there's been about 6+ months of sitting with
some false-starts on some of this and now it's beginning to snap
together so we now go update things like the EdgeTPU support.