Tidy up things a bit.
It helps with WIP stuff being checked in finally if you actually
have the recipe do the install correctly.  Need to flesh out the
path there, but empty stubbed out blocks don't work in Yocto's
"shell" scripting framework because it wants to parse and process
all of it before doing it.