Move to new override syntax
We got a shiny new override syntax for things.  Makes the
declarations, etc. cleaner for build- but at the price of
needing to move eventually.  Now, rather than later since
this is only supporting Dunfell and later where they've
backported the new ROE to and later.  Since we're Gatesgarth
and later, this was the sensible thing to do here.